feat(mergeMapTo): add higher-order lettable version of mergeMapTo
