+++ src/operator/mergeMapTo.ts
@@ -1,11 +1,5 @@
 import { Observable, ObservableInput } from '../Observable';
-import { Operator } from '../Operator';
-import { PartialObserver } from '../Observer';
-import { Subscriber } from '../Subscriber';
-import { Subscription } from '../Subscription';
-import { OuterSubscriber } from '../OuterSubscriber';
-import { InnerSubscriber } from '../InnerSubscriber';
-import { subscribeToResult } from '../util/subscribeToResult';
+import { mergeMapTo as higherOrder } from '../operators/mergeMapTo';
 
 /* tslint:disable:max-line-length */
 export function mergeMapTo<T, R>(this: Observable<T>, observable: ObservableInput<R>, concurrent?: number): Observable<R>;
@@ -58,110 +52,5 @@ export function mergeMapTo<T, I, R>(this: Observable<T>, observable: ObservableI
 export function mergeMapTo<T, I, R>(this: Observable<T>, innerObservable: Observable<I>,
                                     resultSelector?: ((outerValue: T, innerValue: I, outerIndex: number, innerIndex: number) => R) | number,
                                     concurrent: number = Number.POSITIVE_INFINITY): Observable<R> {
-  if (typeof resultSelector === 'number') {
-    concurrent = <number>resultSelector;
-    resultSelector = null;
-  }
-  return this.lift(new MergeMapToOperator(innerObservable, <any>resultSelector, concurrent));
-}
-
-// TODO: Figure out correct signature here: an Operator<Observable<T>, R>
-//       needs to implement call(observer: Subscriber<R>): Subscriber<Observable<T>>
-export class MergeMapToOperator<T, I, R> implements Operator<Observable<T>, R> {
-  constructor(private ish: ObservableInput<I>,
-              private resultSelector?: (outerValue: T, innerValue: I, outerIndex: number, innerIndex: number) => R,
-              private concurrent: number = Number.POSITIVE_INFINITY) {
-  }
-
-  call(observer: Subscriber<R>, source: any): any {
-    return source.subscribe(new MergeMapToSubscriber(observer, this.ish, this.resultSelector, this.concurrent));
-  }
-}
-
-/**
- * We need this JSDoc comment for affecting ESDoc.
- * @ignore
- * @extends {Ignored}
- */
-export class MergeMapToSubscriber<T, I, R> extends OuterSubscriber<T, I> {
-  private hasCompleted: boolean = false;
-  private buffer: T[] = [];
-  private active: number = 0;
-  protected index: number = 0;
-
-  constructor(destination: Subscriber<R>,
-              private ish: ObservableInput<I>,
-              private resultSelector?: (outerValue: T, innerValue: I, outerIndex: number, innerIndex: number) => R,
-              private concurrent: number = Number.POSITIVE_INFINITY) {
-    super(destination);
-  }
-
-  protected _next(value: T): void {
-    if (this.active < this.concurrent) {
-      const resultSelector = this.resultSelector;
-      const index = this.index++;
-      const ish = this.ish;
-      const destination = this.destination;
-
-      this.active++;
-      this._innerSub(ish, destination, resultSelector, value, index);
-    } else {
-      this.buffer.push(value);
-    }
-  }
-
-  private _innerSub(ish: ObservableInput<I>,
-                    destination: PartialObserver<I>,
-                    resultSelector: (outerValue: T, innerValue: I, outerIndex: number, innerIndex: number) => R,
-                    value: T,
-                    index: number): void {
-    this.add(subscribeToResult<T, I>(this, ish, value, index));
-  }
-
-  protected _complete(): void {
-    this.hasCompleted = true;
-    if (this.active === 0 && this.buffer.length === 0) {
-      this.destination.complete();
-    }
-  }
-
-  notifyNext(outerValue: T, innerValue: I,
-             outerIndex: number, innerIndex: number,
-             innerSub: InnerSubscriber<T, I>): void {
-    const { resultSelector, destination } = this;
-    if (resultSelector) {
-      this.trySelectResult(outerValue, innerValue, outerIndex, innerIndex);
-    } else {
-      destination.next(innerValue);
-    }
-  }
-
-  private trySelectResult(outerValue: T, innerValue: I,
-                          outerIndex: number, innerIndex: number): void {
-    const { resultSelector, destination } = this;
-    let result: R;
-    try {
-      result = resultSelector(outerValue, innerValue, outerIndex, innerIndex);
-    } catch (err) {
-      destination.error(err);
-      return;
-    }
-
-    destination.next(result);
-  }
-
-  notifyError(err: any): void {
-    this.destination.error(err);
-  }
-
-  notifyComplete(innerSub: Subscription): void {
-    const buffer = this.buffer;
-    this.remove(innerSub);
-    this.active--;
-    if (buffer.length > 0) {
-      this._next(buffer.shift());
-    } else if (this.active === 0 && this.hasCompleted) {
-      this.destination.complete();
-    }
-  }
+  return higherOrder(innerObservable, resultSelector as any, concurrent)(this);
 }
